修复小数点，开启审核

+++ ruoyi-admin/src/main/java/com/ruoyi/web/controller/custom/SysOrderController.java
@@ -81,7 +81,7 @@ public class SysOrderController extends OrderBaseController {
                 }
             }
         }
-        return getOrderDataTable(list, totalAmount.longValue());
+        return getOrderDataTable(list, totalAmount);
     }
